Some additional fixes to the intermittent failures of the windows tests.
Windows has some serious issues with file and directory removals when they happen rapidly, seemingly in the same folder in two threads, or just in general.  There seem to be a few causes, but primarily that windows will lock a file preventing its removal if it is loaded by a program.  Sometimes this can happen from the intended calling application, or even just an anti-virus.  So there isn't actually a 100% reliable way to delete a file without multiple retries on windows.